% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BCeF_code.R
\name{BCeF}
\alias{BCeF}
\title{A BCeF Function}
\usage{
BCeF(input.edata, input.covariates.df, input.gold.standard,
  input.edata.description = "", input.adjustment.method.description = "")
}
\arguments{
\item{input.edata}{the expression data set to be adjusted.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
BCeF()
}
\keyword{fff}

