% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BCeF_code.R
\name{plot.roc.AUC}
\alias{plot.roc.AUC}
\title{A plot.roc.AUC Function}
\usage{
plot.roc.AUC(gold.standard.bool, predictor.df, main.text, legend.to.use,
  color.to.use)
}
\arguments{
\item{gold.standard.bool}{is a boolean vector representing the the gold standard.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
BCeF()
}
\keyword{fff}

